calories-be/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
httpx
pydantic
python-dotenv
pydantic[email]
pyjwt
passlib
pytz